% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/allocateDelta.R
\name{allocateDelta}
\alias{allocateDelta}
\title{Allocate Delta via Maximum Likelihood}
\usage{
allocateDelta(param1, param2, dist = rep("Normal", length(param1)), sign)
}
\arguments{
\item{param1}{A vector of the first parameter for each of the elements.  For
a normal distribution, this is the mean.}

\item{param2}{A vector of the second parameter for each of the elements. For
a normal distribution, this is the standard deviation.}

\item{dist}{A vector of the name of the distribution for each distribution.
Currently, only "Normal" is implemented.}

\item{sign}{A vector of the sign of each element.  These values should all
  be +1 or -1, and they indicate if Delta_1, Delta_2, ... should be
  pre-multiplied by a negative or not.  Usually, these will all be +1.}
}
\value{
A vector of the final values (Delta_1, Delta_2, ..., Delta_N)
}
\description{
This function assumes that a sum of several elements has been updated and
that the adjustment must be split up into each of the individual elements in
the sum.  Thus, we start with an equation like:
}
\details{
X_1 + X_2 + ... + X_N = Y

Y becomes Y + Delta, and we must determine a new equation such that

(X_1 + Delta_1) + (X_2 + Delta_2) + ... + (X_N + Delta_N) = Y + Delta

The individual delta values should be determined via maximum likelihood, as
we have a distribution for each of these elements.  We can simplify the
problem a bit by subtracting Y from both sides:

Delta_1 + Delta_2 + ... + Delta_N = Delta
}

