% Generated by roxygen2 (4.0.1): do not edit by hand
\name{appendSymbol}
\alias{appendSymbol}
\title{Function to fill symbol based on calculated value}
\usage{
appendSymbol(value, newSymbol, missingSymbol = "M")
}
\arguments{
\item{value}{The vector of calculated values}

\item{newSymbol}{The new symbol to be assigned}

\item{missingSymbol}{The symbol filled in when value is missing}
}
\description{
The function will take a vector of values and fill in the new
symbol if the value is not missing. If the value is missing, it
will insert the missing symbol.
}

