% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Aupus}
\alias{Aupus}
\title{The function performs the whole aupus procedure}
\usage{
Aupus(nodes, edges, from, to, aupusParam)
}
\arguments{
\item{nodes}{The nodes returned in the list from the function
suaToNetworkRepresentation.}

\item{edges}{The edge returned in the list from the function
suaToNetworkRepresentation.}

\item{from}{The column name corresponding to the starting nodes.}

\item{to}{The column corresponding to the target nodes.}
}
\description{
The function performs the whole aupus procedure
}

