% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/balanceProductionElements.R
\name{balanceProductionElements}
\alias{balanceProductionElements}
\title{Balance Production Elements}
\usage{
balanceProductionElements(value31, value41, value51, symb31, symb41, symb51,
  divisionFactor)
}
\arguments{
\item{value31}{A vector of values corresponding to element 31 (Actual
Producing Factor).}

\item{value41}{A vector of values corresponding to element 41 (Productivity
Element).}

\item{value51}{A vector of values corresponding to element 51 (Output).}

\item{symb31}{A vector of sybmols corresponding to element 31.  These are
provided so that they can be returned (updated, if necessary, when a
balance occurs).}

\item{symb41}{A vector of sybmols corresponding to element 41.  These are
provided so that they can be returned (updated, if necessary, when a
balance occurs).}

\item{symb51}{A vector of sybmols corresponding to element 51.  These are
provided so that they can be returned (updated, if necessary, when a
balance occurs).}

\item{divisionFactor}{A numeric vector taking values of 1, 1000, or 10000.
This number is required for the balance equation, and is determined based
on item type.}
}
\value{
A list of 6 elements containing values 31, 41, and 51 and symbols
31, 41, and 51.  All elements will be the same as what was passed except
when a balance was possible, and in this case one of the three values will
be updated along with it's symbol set to "C".
}
\description{
This function takes a data.table and balances the production elements (i.e.
elements 31, 41, and 51).  The relationship is
51 = 31 * 41 / divisionFactor, where divisionFactor is either 1, 1000, or
10000 depending on the item type.
}
\details{
If two elements are available, the third will be computed.  If only one or
zero elements are available, no values will be updated.  divisionFactor
should always be supplied.
}

