% Generated by roxygen2 (4.0.1): do not edit by hand
\name{buildEdges}
\alias{buildEdges}
\title{Function to merge shares and extraction rate and build the
relation graph.}
\usage{
buildEdges(extractionRateData, shareData, inputData)
}
\arguments{
\item{aupusData}{The aupus data from the function getAupus}

\item{extractionRate}{The column name corresponding to extaction rate}

\item{shareData}{The share data from the function getShare}

\item{inputData}{The input data returned by the function getInputData}
}
\description{
Function to merge shares and extraction rate and build the
relation graph.
}

