% Generated by roxygen2 (4.0.1): do not edit by hand
\name{buildNodes}
\alias{buildNodes}
\title{Function to merge aupus, ratio, balance element, item info and
whether to use AUPUS data.}
\usage{
buildNodes(aupusData, ratioData, balanceElementData, itemInfoData,
  balanceElementNum, populationData, verbose = FALSE)
}
\arguments{
\item{aupusData}{The aupus data from the function getAupus}

\item{ratioData}{The ratio data from the function getRatio}

\item{balanceElementData}{The balance element data from the
function getBalanceElement.}

\item{itemInfoData}{The item information data from the function
getItemInfo.}

\item{balanceElementNum}{The column name which corresponds to
balance element.}
}
\description{
Function to merge aupus, ratio, balance element, item info and
whether to use AUPUS data.
}

