% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateAupusElements}
\alias{calculateAupusElements}
\title{This is the function which performs the full aupus module}
\usage{
calculateAupusElements(aupusFinalData, itemTypeCol, balanceElementNum,
  aupusParam)
}
\arguments{
\item{aupusFinalData}{The final data set returned by the mergeAll
function.}

\item{shareData}{The share data from returned from the function
getShareData}

\item{itemTypeCol}{The column corresponding to the item type of
the commodity item.}

\item{shareNum}{The column corresponding to the share in the share
data.}

\item{balanceElementNum}{The column corresponding to the balance
element.}
}
\description{
This is the function which performs the full aupus module
}

