% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateBalance}
\alias{calculateBalance}
\title{The function will calculate the balance and fill in the element to
be balanced.}
\usage{
calculateBalance(supply, utilization, element161Num, element171Num,
  element181Num, element181Symb, balanceElement, itemTypeCol, aupusParam, data)
}
\arguments{
\item{supply}{The column corresponding to calculated total supply.}

\item{utilization}{The column correspoding to calculated total
utilization.}

\item{element161Num}{The column corresponding to value of element
161.}

\item{element171Num}{The column corresponding to value of element
171.}

\item{element181Num}{The column corresponding to value of element
181.}

\item{element181Symb}{The column corresponding to symbol of element
181.}

\item{balanceElement}{The column corresponding to the balancing
element.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
The function calculates the balance based on suppply and
utilization and fill in the balancing element.
}

