% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle101}
\alias{calculateEle101}
\title{The function calculates element 101 (Use for animal)}
\usage{
calculateEle101(element101Num, element101Symb, ratio101Num, stotal, data)
}
\arguments{
\item{element101Num}{The column corresponding to value of element
101.}

\item{element101Symb}{The column corresponding to symbol of
element 101.}

\item{ratio101Num}{The column corresponding to the ratio of
element 101.}

\item{stotal}{The column corresponding to the total supply.}

\item{data}{The data}
}
\description{
The function calculates element 101 (Use for animal)
}

