% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle11}
\alias{calculateEle11}
\title{Function to calculate element 11 (Initial existence)}
\usage{
calculateEle11(element11Num, element11Symb, element161Num, element161Symb,
  itemTypeCol, data)
}
\arguments{
\item{element11Num}{The column corresponding to value of element
11.}

\item{element11Symb}{The column corresponding to symbol of element
11.}

\item{element161Num}{The column corresponding to value of element
161.}

\item{element161Symb}{The column corresponding to symbol of element
161.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
Function to calculate element 11 (Initial existence)
}

