% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle111}
\alias{calculateEle111}
\title{The function calculates element 111 (Use for same product)}
\usage{
calculateEle111(element111Num, element111Symb, element21Num, element31Num,
  ratio171Num, ratio111Num, stotal, data)
}
\arguments{
\item{element111Num}{The column corresponding to value of element
111.}

\item{element111Symb}{The column corresponding to symbol of element
111.}

\item{element21Num}{The column corresponding to value of element
21.}

\item{element31Num}{The column corresponding to value of element
31.}

\item{ratio171Num}{The column corresponding to ratio of element
171.}

\item{ratio111Num}{The column corresponding to ratio of element
111.}

\item{stotal}{The column corresponding to the total supply.}

\item{data}{The data}
}
\description{
The function calculates element 111 (Use for same product)
}

