% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle121}
\alias{calculateEle121}
\title{The function to calculate element 121 (Losses)}
\usage{
calculateEle121(element121Num, element121Symb, ratio121Num, stotal, data)
}
\arguments{
\item{element121Num}{The column corresponding to value of element
121.}

\item{element121Symb}{The column corresponding to symbol of element
121.}

\item{ratio121Num}{The column corresponding to ratio of element
121.}

\item{stotal}{The column corresponding to total supply}

\item{data}{The data}
}
\description{
The function to calculate element 121 (Losses)
}

