% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle131}
\alias{calculateEle131}
\title{The function calculates element 131 (Reemployment Same Sector)}
\usage{
calculateEle131(element131Num, element131Symb, ratio131Num, stotal, data)
}
\arguments{
\item{element131Num}{The column corresponding to value of element
131.}

\item{element131Symb}{The column corresponding to symbol of element
131.}

\item{ratio131Num}{The column corresponding to ratio of element
131.}

\item{stotal}{The column corresponding to total supply.}

\item{data}{The data}
}
\description{
The function calculates element 131 (Reemployment Same Sector)
}

