% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle141}
\alias{calculateEle141}
\title{The function calculates element 141 (Consumption)}
\usage{
calculateEle141(element141Num, element141Symb, element11Num, element51Num,
  element61Num, element91Num, element95Num, element161Num, ratio141Num, stotal,
  itemTypeCol, data)
}
\arguments{
\item{element141Num}{The column corresponding to value of element
141.}

\item{element141Symb}{The column corresponding to symbol of element
141.}

\item{element11Num}{The column corresponding to value of element
11.}

\item{element51Num}{The column corresponding to value of element
51.}

\item{element61Num}{The column corresponding to value of element
61.}

\item{element91Num}{The column corresponding to value of element
91.}

\item{element95Num}{The column corresponding to value of element
95.}

\item{element161Num}{The column corresponding to value of element
161.}

\item{ratio141Num}{The column corresponding to ratio of element
141.}

\item{stotal}{The column corresponding to total supply.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
The function calculates element 141 (Consumption)
}

