% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle144}
\alias{calculateEle144}
\title{The function calculates element 144 (consumption per day)}
\usage{
calculateEle144(element144Num, element144Symb, element141Num, population11Num,
  itemTypeCol, data)
}
\arguments{
\item{element144Num}{The column corresponding to value of element
144.}

\item{element144Symb}{The column corresponding to symbol of element
144.}

\item{element141Num}{The column corresponding to value of element
141.}

\item{population11Num}{The column corresponds to element 11 of
population.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
The function calculates element 144 (consumption per day)
}

