% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle151}
\alias{calculateEle151}
\title{The function calculates element 151 (Reemployment other sector)}
\usage{
calculateEle151(element151Num, element151Symb, element131Num, element51Num,
  ratio151Num, stotal, data)
}
\arguments{
\item{element151Num}{The column corresponding to value of element
151.}

\item{element151Symb}{The column corresponding to symbol of element
151.}

\item{element131Num}{The column corresponding to value of element
131.}

\item{element51Num}{The column corresponding to value of element
51.}

\item{ratio151Num}{The column corresponding to ratio of element
151.}

\item{stotal}{The column corresponding to total supply.}

\item{data}{The data}
}
\description{
The function calculates element 151 (Reemployment other sector)
}

