% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle161}
\alias{calculateEle161}
\title{The function calculates element 161 (Final existence)}
\usage{
calculateEle161(element161Num, element161Symb, element11Num, element71Num,
  itemTypeCol, data)
}
\arguments{
\item{element161Num}{The column corresponding to value of element
161.}

\item{element161Symb}{The column corresponding to symbol of element
161.}

\item{element11Num}{The column corresponding to value of element
11.}

\item{element71Num}{The column corresponding to value of element
71.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
The function calculates element 161 (Final existence)
}

