% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle171}
\alias{calculateEle171}
\title{The function calculates element 171 (consumption sugar)}
\usage{
calculateEle171(element171Num, element171Symb, element101Num, element121Num,
  element131Num, element141Num, element151Num, data)
}
\arguments{
\item{element171Num}{The column corresponding to value of element
171.}

\item{element171Symb}{The column corresponding to symbol of element
171.}

\item{element101Num}{The column corresponding to value of element
101.}

\item{element121Num}{The column corresponding to value of element
121.}

\item{element131Num}{The column corresponding to value of element
131.}

\item{element141Num}{The column corresponding to value of element
141.}

\item{element151Num}{The column corresponding to value of element
151.}

\item{data}{The data}
}
\description{
The function calculates element 171 (consumption sugar)
}

