% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle174}
\alias{calculateEle174}
\title{The function calculates element 174 (consumption per caput sugar)}
\usage{
calculateEle174(element174Num, element174Symb, element171Num, population11Num,
  data)
}
\arguments{
\item{element174Num}{The column corresponding to value of element
174.}

\item{element174Symb}{The column corresponding to symbol of element
174.}

\item{element171Num}{The column corresponding to value of element
171.}

\item{population11Num}{The column corresponds to element 11 of the
population.}

\item{data}{The data}
}
\description{
The function calculates element 174 (consumption per caput sugar)
}

