% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle21}
\alias{calculateEle21}
\title{This function calculates element 21 (potential producing factor)}
\usage{
calculateEle21(element21Num, element21Symb, element11Num, element111Num,
  ratio171Num, itemTypeCol, data)
}
\arguments{
\item{element21Num}{The column corresponding to value of element
21.}

\item{element21Symb}{The column corresponding to symbol of element
21.}

\item{element111Num}{The column corresponding to value of element
111.}

\item{ratio171Num}{The column corresponding to ratio of element
171.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function calculates element 21 (potential producing factor)
}

