% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle261}
\alias{calculateEle261}
\title{This function calculates element 261 (calories)}
\usage{
calculateEle261(element261Num, element261Symb, ratio261Num, element141Num, data)
}
\arguments{
\item{element261Num}{The column corresponding to value of element
261.}

\item{element261Symb}{The column corresponding to symbol of element
261.}

\item{ratio261Num}{The column corresponding to ratio of element
261.}

\item{element141Num}{The column corresponding to value of element
141.}

\item{data}{The data}
}
\description{
This function calculates element 261 (calories)
}

