% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle264}
\alias{calculateEle264}
\title{This function calculates element 264 (calorie per day)}
\usage{
calculateEle264(element264Num, element264Symb, element261Num, population11Num,
  population21Num, data)
}
\arguments{
\item{element264Num}{The column corresponding to value of element
264.}

\item{element264Symb}{The column corresponding to symbol of element
264.}

\item{element261Num}{The column corresponding to value of element
261.}

\item{population11Num}{The column corresponds to element 11 of the
population.}

\item{population21Num}{The column corresponds to element 21 of the
population.}

\item{data}{The data}
}
\description{
This function calculates element 264 (calorie per day)
}

