% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle271}
\alias{calculateEle271}
\title{This function calculates element 271 (protein)}
\usage{
calculateEle271(element271Num, element271Symb, ratio271Num, element141Num, data)
}
\arguments{
\item{element271Num}{The column corresponding to value of element
271.}

\item{element271Symb}{The column corresponding to symbol of element
271.}

\item{ratio271Num}{The column corresponding to ratio of element
271.}

\item{element141Num}{The column corresponding to value of element
141.}

\item{data}{The data}
}
\description{
This function calculates element 271 (protein)
}

