% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle274}
\alias{calculateEle274}
\title{This function calculates element 274 (protein per day)}
\usage{
calculateEle274(element274Num, element274Symb, element261Num, population11,
  population21, data)
}
\arguments{
\item{element274Num}{The column corresponding to value of element
274.}

\item{element274Symb}{The column corresponding to symbol of element
274.}

\item{element261Num}{The column corresponding to value of element
261.}

\item{population11Num}{The column corresponds to element 11 of the
population.}

\item{population21Num}{The column corresponds to element 21 of the
population.}

\item{data}{The data}
}
\description{
This function calculates element 274 (protein per day)
}

