% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle281}
\alias{calculateEle281}
\title{This function calculates element 281 (fat)}
\usage{
calculateEle281(element281Num, element281Symb, ratio281Num, element141Num, data)
}
\arguments{
\item{element281Num}{The column corresponding to value of element
281.}

\item{element281Symb}{The column corresponding to symbol of element
281.}

\item{ratio281Num}{The column corresponding to ratio of element
281.}

\item{element141Num}{The column corresponding to value of element
141.}

\item{data}{The data}
}
\description{
This function calculates element 281 (fat)
}

