% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle284}
\alias{calculateEle284}
\title{This function calculates element 284 (fat per day)}
\usage{
calculateEle284(element284Num, element284Symb, element261Num, population11,
  population21, data)
}
\arguments{
\item{element284Num}{The column corresponding to value of element
284.}

\item{element284Symb}{The column corresponding to symbol of element
284.}

\item{element261Num}{The column corresponding to value of element
261.}

\item{population11Num}{The column corresponds to element 11 of the
population.}

\item{population21Num}{The column corresponds to element 21 of the
population.}

\item{data}{The data}
}
\description{
This function calculates element 284 (fat per day)
}

