% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle31}
\alias{calculateEle31}
\title{This function calculates element 31}
\usage{
calculateEle31(element31Num, element31Symb, inputNum, itemTypeCol, data)
}
\arguments{
\item{element31Num}{The column corresponding to value of element
31.}

\item{element31Symb}{The column corresponding to symbol of element
31.}

\item{inputNum}{The column corresponds to input from processing}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function calculates element 31
}

