% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle314151}
\alias{calculateEle314151}
\title{This function balance and fill in missing element 31, 41, and 51.}
\usage{
calculateEle314151(element31Num, element41Num, element51Num, element31Symb,
  element41Symb, element51Symb, itemTypeCol, data)
}
\arguments{
\item{element31Num}{The column corresponding to value of element
31.}

\item{element31Symb}{The column corresponding to symbol of element
31.}

\item{element41Num}{The column corresponding to value of element
41.}

\item{element41Symb}{The column corresponding to symbol of element
41.}

\item{element51Num}{The column corresponding to value of element
51.}

\item{element51Symb}{The column corresponding to symbol of element
51.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function balance and fill in missing element 31, 41, and 51.
}

