% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle41}
\alias{calculateEle41}
\title{The function calculates element 41}
\usage{
calculateEle41(element41Num, element41Symb, ratio41Num, data)
}
\arguments{
\item{element41Num}{The column corresponding to value of element
41.}

\item{element41Symb}{The column corresponding to symbol of element
41.}

\item{ratio41Num}{The column corresponding to ratio of element
41.}

\item{data}{The data}
}
\description{
The function calculates element 41
}

