% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle51}
\alias{calculateEle51}
\title{The function calculates element 51}
\usage{
calculateEle51(element51Num, element51Symb, element58Num, itemTypeCol, data)
}
\arguments{
\item{element51Num}{The column corresponding to value of element
51.}

\item{element51Symb}{The column corresponding to symbol of element
51.}

\item{element58Num}{The column corresponding to value of element
58.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
The function calculates element 51
}

