% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle541}
\alias{calculateEle541}
\title{This function calculates element 541 (final demand)}
\usage{
calculateEle541(element541Num, element541Symb, element542Num, element543Num,
  element544Num, element545Num, data)
}
\arguments{
\item{element541Num}{The column corresponding to value of element
541.}

\item{element541Symb}{The column corresponding to symbol of element
541.}

\item{element542Num}{The column corresponding to value of element
542.}

\item{element543Num}{The column corresponding to value of element
543.}

\item{element544Num}{The column corresponding to value of element
544.}

\item{element545Num}{The column corresponding to value of element
545.}

\item{data}{The data}
}
\description{
This function calculates element 541 (final demand)
}

