% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle546}
\alias{calculateEle546}
\title{This function calculates element 546}
\usage{
calculateEle546(element546Num, element546Symb, element541Num, element151Num,
  element191Num, data)
}
\arguments{
\item{element546Num}{The column corresponding to value of element
546.}

\item{element546Symb}{The column corresponding to symbol of element
546.}

\item{element541Num}{The column corresponding to value of element
541.}

\item{element151Num}{The column corresponding to value of element
151.}

\item{element191Num}{The column corresponding to value of element
191.}

\item{data}{The data}
}
\description{
This function calculates element 546
}

