% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle58}
\alias{calculateEle58}
\title{This function calculates element 58 (production crop year)}
\usage{
calculateEle58(element58Num, element58Symb, itemTypeCol, data)
}
\arguments{
\item{element58Num}{The column corresponding to value of element
58.}

\item{element58Symb}{The column corresponding to symbol of element
58.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function calculates element 58 (production crop year)
}

