% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle63}
\alias{calculateEle63}
\title{This function calculates element 63 (inflow unit value)}
\usage{
calculateEle63(element61Num, element62Num, element63Num, element63Symb, data)
}
\arguments{
\item{element61Num}{The column corresponding to value of element
61.}

\item{element62Num}{The column corresponding to value of element
62.}

\item{element63Num}{The column corresponding to value of element
63.}

\item{element63Symb}{The column corresponding to symbol of element
63.}

\item{data}{The data}
}
\description{
This function calculates element 63 (inflow unit value)
}

