% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle66}
\alias{calculateEle66}
\title{This function calculates element 66 (standardized inflow)}
\usage{
calculateEle66(element41Num, element61Num, element66Num, element66Symb, shares,
  itemTypeCol, data, shareData)
}
\arguments{
\item{element41Num}{The column corresponding to value of element
41.}

\item{element61Num}{The column corresponding to value of element
61.}

\item{element66Num}{The column corresponding to value of element
66.}

\item{element66Symb}{The column corresponding to symbol of element
66.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function calculates element 66 (standardized inflow)
}

