% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle71}
\alias{calculateEle71}
\title{This function calculates element 71 (from initial existence)}
\usage{
calculateEle71(element71Num, element71Symb, element51Num, element61Num,
  element91Num, element101Num, element121Num, element131Num, element141Num,
  element151Num, element161Num, itemTypeCol, data)
}
\arguments{
\item{element71Num}{The column corresponding to value of element
71.}

\item{element71Symb}{The column corresponding to symbol of element
71.}

\item{element51Num}{The column corresponding to value of element
51.}

\item{element61Num}{The column corresponding to value of element
61.}

\item{element91Num}{The column corresponding to value of element
91.}

\item{element101Num}{The column corresponding to value of element
101.}

\item{element121Num}{The column corresponding to value of element
121.}

\item{element131Num}{The column corresponding to value of element
131.}

\item{element141Num}{The column corresponding to value of element
141.}

\item{element151Num}{The column corresponding to value of element
151.}

\item{element161Num}{The column corresponding to value of element
161.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function calculates element 71 (from initial existence)
}

