% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle93}
\alias{calculateEle93}
\title{This function calculates element 93 (outflow unit value)}
\usage{
calculateEle93(element91Num, element92Num, element93Num, element93Symb,
  itemTypeCol, data)
}
\arguments{
\item{element91Num}{The column corresponding to value of element
91.}

\item{element92Num}{The column corresponding to value of element
92.}

\item{element93Num}{The column corresponding to value of element
93.}

\item{element93Symb}{The column corresponding to symbol of element
93.}

\item{itemTypeCol}{The column corresponding to the type of the
commodity item.}

\item{data}{The data}
}
\description{
This function calculates element 93 (outflow unit value)
}

