% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateEle96}
\alias{calculateEle96}
\title{This function calculates element 99 (standardized outflow)}
\usage{
calculateEle96(element41Num, element91Num, element96Num, element96Symb, shares,
  itemTypeCol, data, shareData)
}
\arguments{
\item{element41Num}{The column corresponding to value of element
41.}

\item{element91Num}{The column corresponding to value of element
91.}

\item{element96Num}{The column corresponding to value of element
96.}

\item{element96Symb}{The column corresponding to symbol of element
96.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function calculates element 99 (standardized outflow)
}

