% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateTotalInput}
\alias{calculateTotalInput}
\title{This function calculates total input}
\usage{
calculateTotalInput(inputData, inputNum, aupusParam)
}
\arguments{
\item{inputData}{The data containing the input extracted from the
function getInputFromProcess after running the
updateInputFromProcess function.}
}
\description{
This function calculates total input
}

