% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateTotalNutritive.R
\name{calculateTotalNutritive}
\alias{calculateTotalNutritive}
\title{Calculate Total Nutritive (Calories, Proteins, Fats)}
\usage{
calculateTotalNutritive(ratioNum, elementNum, data, aupusParam)
}
\arguments{
\item{ratioNum}{The column name of data corresponding to the ratio for this
element.  Typically one of "Ratio_measuredElementFS_261" for calories,
"Ratio_measuredElementFS_271" for proteins, or
"Ratio_measuredElementFS_261" for fats.}

\item{elementNum}{The element number.  Currently, this should be one of 261
(calories), 271 (proteins) or 281 (fats).}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 261/271/281's value
and symbol updated.
}
\description{
Each commodity has a ratio defined, R_261/R_271/R_281, which gives the
conversion rate of the quantity of the commodity in question into
calories/proteins/fats.  Thus, element 261/271/281 is computed simply as
E_261 = E_141 * R_261 / 100
}
\details{
E_271 = E_141 * R_271 / 1000

E_281 = E_141 * R_281 / 1000
}

