% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateTotalSupply}
\alias{calculateTotalSupply}
\title{This function calculates total supply}
\usage{
calculateTotalSupply(element11Num, element51Num, element58Num, element61Num,
  element66Num, itemTypeCol, data)
}
\arguments{
\item{element11Num}{The column corresponding to value of element
11.}

\item{element51Num}{The column corresponding to value of element
51.}

\item{element58Num}{The column corresponding to value of element
58.}

\item{element61Num}{The column corresponding to value of element
61.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function calculates total supply
}

