% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calculateTotalUtilization}
\alias{calculateTotalUtilization}
\title{This function calculates total utilization}
\usage{
calculateTotalUtilization(element91Num, element95Num, element96Num,
  element101Num, element111Num, element121Num, element131Num, element141Num,
  element151Num, element161Num, element546Num, itemTypeCol, data)
}
\arguments{
\item{element91Num}{The column corresponding to value of element
91.}

\item{element95Num}{The column corresponding to value of element
95.}

\item{element96Num}{The column corresponding to value of element
96.}

\item{element101Num}{The column corresponding to value of element
101.}

\item{element111Num}{The column corresponding to value of element
111.}

\item{element121Num}{The column corresponding to value of element
121.}

\item{element131Num}{The column corresponding to value of element
131.}

\item{element141Num}{The column corresponding to value of element
141.}

\item{element151Num}{The column corresponding to value of element
151.}

\item{element161Num}{The column corresponding to value of element
161.}

\item{element546Num}{The column corresponding to value of element
546.}

\item{itemTypeCol}{The column which identifies the item type of
the commodity item.}

\item{data}{The data}
}
\description{
This function calculates total utilization
}

