% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkShareUnity}
\alias{checkShareUnity}
\title{This is a function checking whether the shares sums up to 100.}
\usage{
checkShareUnity(shareData)
}
\arguments{
\item{shareData}{The share data returned from getShare.}
}
\description{
This is a function checking whether the shares sums up to 100.
}

