% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coerceColumnType.R
\name{coerceColumnTypes}
\alias{coerceColumnTypes}
\title{Coerce Column Types}
\usage{
coerceColumnTypes(aupusParam, data)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{data}{The data.table whose column types should be coerced.}
}
\value{
A data.table with updated column types.
}
\description{
Currently, the working system may not provide the correct column types for
datasets, and this can cause problems during joins at later stages (as keys
are sometimes numeric and sometimes character, and a failure to match leads
to errors).  Thus, this function coerces all column types that it can,
based on the aupusParam argument.  For example,
aupusParam$keyNames$areaName gives the name of the area id column of data,
and thus this column should always be coerced to a numeric column.
}
\details{
Note: there is currently a ticket in the SWS (issues SWS-797) that should
resolve this problem.  However, we need a work-around until that issue has
been resolved.
}

