% Generated by roxygen2 (4.0.1): do not edit by hand
\name{collapseShare}
\alias{collapseShare}
\title{Function to collapse the share data}
\usage{
collapseShare(shareData, shares, aupusParam, verbose = FALSE)
}
\arguments{
\item{shareData}{The shares data obtained from the function
getShare.}

\item{shares}{The column correspond to shares.}

\item{verbos}{Whether the output should be printed.}
}
\description{
Function to collapse the share data
}

