% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeProcessedProduction.R
\name{computeProcessedProduction}
\alias{computeProcessedProduction}
\title{Compute Processed Production}
\usage{
computeProcessedProduction(data, tree, standParams)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{tree}{The commodity tree which provides the edge structure.}

\item{standParams}{The parameters for standardization.  These parameters
  provide information about the columns of data and tree, specifying (for
  example) which columns should be standardized, which columns represent
  parents/children, etc.}
}
\value{
Nothing is returned, but the passed data.table "data" is updated
  with estimated distributions for the production for all elements which did
  not previously have distributions.
}
\description{
This function estimates the required "production" values for processed
products.  Of course, processed products don't really have true "production"
values, but this rather represents how much of the current commodity is
available because of processing of a parent commodity.  The estimation of
the required amount of production is done by examining the imbalance between
imports, exports, and food (i.e. production = exports + food - imports, or 0
if imports > exports + food).  This value is then standardized to provide an
estimated distribution for the parent product.
}

