% Generated by roxygen2 (4.0.1): do not edit by hand
\name{constructStandardizationGraph}
\alias{constructStandardizationGraph}
\title{This function constructs the graph/network required for standardization}
\usage{
constructStandardizationGraph(nodes, edges, standardizeElement, from, to,
  aupusParam)
}
\arguments{
\item{nodes}{The nodes of the network representation}

\item{edges}{The edges of the network representation}

\item{standardizeElement}{The names of the elements to be standardized.}

\item{from}{The name corresponding to the from node in the edge.}

\item{to}{The name corresponding to the target node in the edge.}
}
\description{
This function constructs the graph/network required for standardization
}

