% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/elementCodeDescription.R
\docType{data}
\name{elementCodeDescription}
\alias{elementCodeDescription}
\title{Element Code Description}
\format{A data.table}
\usage{
data(elementCodeDescription)
}
\description{
The AUPUS system defines "elements" and stores information in these
variables.  The element code defines what the variable holds, and this
dataset describes which codes correspond to what variables.  The first two
columns are likely to be the most useful, as they contain the code and a
short description.  The remaining columns give examples of what that
element code/variable represents for certain commodity groups.
}
\keyword{aupus}
\keyword{code,}
\keyword{element}

