% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fbsStandardization}
\alias{fbsStandardization}
\title{This function takes the graph and perform the standardization}
\usage{
fbsStandardization(graph, standardizeElement, aupusParam, plot)
}
\arguments{
\item{graph}{The standardization graph from the function
constructStandardizationGraph.}

\item{standardizeElement}{The node sttributes to be standardized.}

\item{plot}{Whether the graph/network should be plotted}
}
\description{
This function takes the graph and perform the standardization
}

