% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fillBalance}
\alias{fillBalance}
\title{This function returns fills the balance}
\usage{
fillBalance(calculatedBalance, balanceNum, balanceSymb, element181Num,
  element181Symb, data)
}
\arguments{
\item{calculatedBalance}{The column corresponding to the balance
calculated by the function calculateBalance.}

\item{balanceNum}{The value column corresponding to the balancing
element.}

\item{balanceSymb}{The symb column corresponding to the balancing
element.}

\item{element181Num}{The column corresponding to value of element
181.}

\item{element181Symb}{The column corresponding to symbol of element
181.}

\item{data}{The data}
}
\description{
The balance calculated from the calculate balance function is
filled in either the balance element or in discrepancy element
181.
}

