% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fillMissingColumn}
\alias{fillMissingColumn}
\title{This function creates columns which are missing.}
\usage{
fillMissingColumn(data, allColumn)
}
\arguments{
\item{data}{The data}

\item{allColumn}{The theoretical set of all column names.}
}
\description{
Some countries may not have a certain element calculated and thus
when the data is extracted, the column is missing. This function
will recreate the missing column.
}

