% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findConnectedGraph}
\alias{findConnectedGraph}
\title{This function takes a graph/network and then find the subset of
the network which are connected to the specified nodes/commodities.}
\usage{
findConnectedGraph(graph, commodity)
}
\arguments{
\item{graph}{The graph object created by the function
constrctGraph.}

\item{commodity}{The commodities}
}
\description{
This function takes a graph/network and then find the subset of
the network which are connected to the specified nodes/commodities.
}

