% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findProcessingLevel}
\alias{findProcessingLevel}
\title{The function finds the processing level of the item in relation
with the rest of the commodity items.}
\usage{
findProcessingLevel(edgeData, from, to, aupusParam, plot = FALSE)
}
\arguments{
\item{edgeData}{the edge data from the function buildEdgeRelation}

\item{from}{The column corresponding to the from node.}

\item{to}{The column corresponding to target node}

\item{plot}{Whether the graph should be plotted.}
}
\description{
The function finds the processing level of the item in relation
with the rest of the commodity items.
}

