% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAupusDataset}
\alias{getAupusDataset}
\title{Function to obtain all required data for the AUPUS module}
\usage{
getAupusDataset(assignGlobal = TRUE, aupusParam)
}
\arguments{
\item{assignGlobal}{Whether the data should be assigned globally
or returned as a list. Default to TRUE, and assigned globally.}
}
\description{
Function to obtain all required data for the AUPUS module
}

