% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getAupusParameter}
\alias{getAupusParameter}
\title{This function gets all the aupusParameter in order to query the data}
\usage{
getAupusParameter(areaCode, assignGlobal = TRUE)
}
\arguments{
\item{areaCode}{The country code of the country of interest}

\item{assignGlobal}{logical, default to FALSE, where a list is
returned. If TRUE, then the result will be assigned globally.}
}
\description{
This function gets all the aupusParameter in order to query the data
}

