% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getExtractionRateData}
\alias{getExtractionRateData}
\title{This function extracts the extraction rate data from the aupus data}
\usage{
getExtractionRateData(database = c("new", "old"), conn, aupusParam)
}
\arguments{
\item{database}{Whether to use the new or the old statistical
working system.}

\item{conn}{The RJDBS connection to the old working system.}
}
\description{
This function extracts the extraction rate data from the aupus data
}

