% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getItemInfoData.R
\name{getItemInfoData}
\alias{getItemInfoData}
\title{Get Item Information Data}
\usage{
getItemInfoData()
}
\value{
A data.table with the codes for each measured item as well as a
name describing the item and an item type.
}
\description{
The name and type of the commodity corresponding to the item code
is returned.
}

