% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPopulationData}
\alias{getPopulationData}
\title{This is the function to obtain the population separately.}
\usage{
getPopulationData(database = c("new", "old"), conn, aupusParam)
}
\arguments{
\item{database}{Whether to use the new or the old statistical
working system.}

\item{conn}{The RJDBS connection to the old working system.}
}
\description{
This is the function to obtain the population separately.
}

