% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSpecificTree.R
\name{getSpecificTree}
\alias{getSpecificTree}
\title{Get Specific Tree}
\usage{
getSpecificTree(aupusData, suaTree)
}
\arguments{
\item{aupusData}{A list of two data.table objects, the names of which should
be "nodes" and "edges".  This object thus contains the AUPUS data.  This
object is usually the output from the Aupus function.}

\item{suaTree}{A data.table containing columns childID, parentID,
  extractionRate, and calorieExtractionRate.  This table specifies how SUA
  level commodities will get standardized to their primary equivalents.}
}
\value{
A data.table with the extraction rates and shares for each
  year/country/parent/child.
}
\description{
This function gets the standardization tree specific to the passed AUPUS
data: it uses the country specific extraction weights when available, and it
uses the standardization shares based on production.  If these specific
values are not available, it uses the default values.
}

