% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getStandardizationTree.R
\name{getStandardizationTree}
\alias{getStandardizationTree}
\title{Get Standardization Tree}
\usage{
getStandardizationTree(aupusER, defaultOnly = FALSE)
}
\arguments{
\item{aupusER}{The extraction rate table from the AUPUS data object.  This
argument is needed because it provides a way to overwrite the default
extraction rates with country specific ones, if available.}

\item{defaultOnly}{Logical.  Should only default extraction rates be used?
  Or should country-specific rates be used when available and default rates
  used in all other cases?  The old methodology isn't clear about what is
  used.}
}
\value{
A data.table object containing 4 columns: child, parent,
  extractionRate, and caloriesOnly.  The child and parent columns contain
  character values of the items, extractionRate contains the extraction
  rates, and caloriesOnly is a logical indicating if the edge of interest
  corresponds to an aggregation of just calories or both calories and
  quantities.
}
\description{
This function pulls the tree for standardization.
}

