% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/itemTypeDescription.R
\docType{data}
\name{itemTypeDescription}
\alias{itemTypeDescription}
\title{Item Type Description}
\format{A data.table with two columns: the itemTypeCode and the
corresponding name.}
\usage{
data(itemTypeDescription)
}
\description{
Each commodity of interest has an "item type".  This item type is a
general grouping of items, and describes what kind of item a particular
commodity is.  Each commodity has an item code, and the item type can
be thought of as defining logical groups of item codes.
}
\keyword{commodity,}
\keyword{group}
\keyword{item}
\keyword{item,}

