% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mergeBalanceElement}
\alias{mergeBalanceElement}
\title{This function appends balance elements to the aupus data}
\usage{
mergeBalanceElement(aupus, balanceElementData, balanceElementNum,
  verbose = FALSE)
}
\arguments{
\item{aupus}{The aupus data extracted from getAupus function.}

\item{balanceElementData}{The balance element data extracted from
getBalanceElement function.}

\item{balanceElementNum}{The column name corresponds to the
balance element values.}

\item{verbose}{To print out information}
}
\description{
The balance elements contains wild card, the function will fill in
the country and year specific values, then year wild card then
global wild card.
}

