% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mergeItemInfo}
\alias{mergeItemInfo}
\title{Function to merge item info to the data}
\usage{
mergeItemInfo(aupusData, itemInfoData)
}
\arguments{
\item{aupusData}{Data returned from the function getAupusData}

\item{itemInfoData}{Data returned from the function
getItemInfoData.}
}
\description{
Function to merge item info to the data
}

