% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mergeRatio}
\alias{mergeRatio}
\title{This function appends ratios to the aupus data}
\usage{
mergeRatio(aupusData, ratioData, verbose = FALSE)
}
\arguments{
\item{aupusData}{The aupus data extracted from getAupus function.}

\item{ratioData}{The ratio data from the function getRatio.}

\item{verbose}{To print out information}
}
\description{
The ratio data contains wild card, the function will fill in
the country and year specific values, then year wild card then
global wild card.
}

