% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numberOfMissingElement}
\alias{numberOfMissingElement}
\title{Function to find the number of elements which are missing from
multiple elements.}
\usage{
numberOfMissingElement(...)
}
\arguments{
\item{...}{The columns of the multiple elements, see base::sum}
}
\description{
Function to find the number of elements which are missing from
multiple elements.
}

