% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numberOfTrendingElement}
\alias{numberOfTrendingElement}
\title{Function to find the number of elements which are trended from
multiple elements.}
\usage{
numberOfTrendingElement(...)
}
\arguments{
\item{...}{The columns of the multiple elements, see base::sum.}
}
\description{
Function to find the number of elements which are trended from
multiple elements.
}

