% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCommodityTree.R
\name{plotCommodityTree}
\alias{plotCommodityTree}
\title{Plot Commodity Tree}
\usage{
plotCommodityTree(shareData, parentName = "measuredItemParentFS",
  childName = "measuredItemChildFS", ...)
}
\arguments{
\item{shareData}{The data.table containing the share information, typically
as generated by getShareData.}

\item{parentName}{The column name of shareData which contains the item code
of the parent.}

\item{childName}{The column name of shareData which contains the item code
of the child.}

\item{...}{Additional parameters to be passed to the plot.igraph function.}
}
\value{
No data is returned, but a plot of the commodity tree is drawn
using the igraph package.
}
\description{
This function is useful for visualizing the relationships between different
commodities as defined within the shareData dataset (see getShareData).
}

