% Generated by roxygen2 (4.0.1): do not edit by hand
\name{replaceable}
\alias{replaceable}
\title{Returns the index which the values can be replaced}
\usage{
replaceable(symb, replaceableSymb = c("C", "T", "M"))
}
\arguments{
\item{symb}{The symbol column correspoinding to the element to be
replaced.}

\item{replaceableSymb}{The original symbol which can be replaced.}
}
\description{
Only data with replaceable symbols are replaced by the
module. This function will return the index in which the values
can be replaced given the corresponding symbol.
}

