% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollDownFoodDelta.R
\name{rollDownFoodDelta}
\alias{rollDownFoodDelta}
\title{Roll Down Food Delta}
\usage{
rollDownFoodDelta(data, tree, standParams, shares, delta)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{tree}{The commodity tree which provides the edge structure.}

\item{standParams}{The parameters for standardization.  These parameters
provide information about the columns of data and tree, specifying (for
example) which columns should be standardized, which columns represent
parents/children, etc.}

\item{shares}{A data.table containing, for each parent commodity, the
allocation of it's processed values into it's children.  Thus, the shares
should sum to one when the sum is done within each parent.}

\item{delta}{The change in the food quantity due to balancing.  A positive
  delta value implies that balancing the food increased it (over the
  original food value).}
}
\value{
A data.table
}
\description{
Balancing of the SUAs is down at the primary product level.  This balancing
will likely adjust the food for that primary commodity, and an adjustment of
the food value thus implies that the production of the children commodities
should be likewise adjusted.  A negative change implies that production of
children commodities should be reduced according to their variance.  A
positive change implies that no adjustment is required, but it is still
important to increase flour production, for example, so that the correct
bran and germ amounts are created.  Thus, production for first processing
level commodities only should be increased if the delta value is positive,
and these increases should be such that the production of the first
processing level products follows the share ratios as closely as possible.
}

