% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollUpFood.R
\name{rollUpFood}
\alias{rollUpFood}
\title{Roll up food}
\usage{
rollUpFood(data, tree, standParams)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{tree}{The commodity tree which provides the edge structure.}

\item{standParams}{The parameters for standardization.  These parameters
  provide information about the columns of data and tree, specifying (for
  example) which columns should be standardized, which columns represent
  parents/children, etc.}
}
\value{
The same data.table as was passed ("data") but with updated food
  distributions to account for standardized production of children
  commodities.
}
\description{
This function takes SUA level data and standardizes up the required
production values for processed commodities into a food value for their
parents.  The idea is that we have official and estimated data at the SUA
level, and the imbalances at this level translate into an imbalance in their
parent commodities.  To capture/account for this imbalance, we roll it up
into food at the top line.  Note that the original estimates for food will
be stored in this table under a new element code: the current food element
code pasted with "_orig".  These original values need to be saved somewhere,
and this seemed like a reasonable place.
}

