% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/saveAupusData.R
\name{SaveAupusData}
\alias{SaveAupusData}
\title{Save AUPUS Data}
\usage{
SaveAupusData(aupusData, nodes)
}
\arguments{
\item{aupusData}{A data.table containing the AUPUS data read from the
database, see ?getAupusDataset.  This is provided solely for extraction
of the appropriate column names, it isn't used in any other way.}

\item{nodes}{The data after having been processed by the AUPUS
proceedure.  This is the data written back to the SWS.}
}
\value{
No data is returned, but the nodes data.set is written back to
the SWS.
}
\description{
This function is designed as a wrapper around SaveData, and is specifically
used to save data produced from the AUPUS analysis back to the SWS.
}

