% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SaveInputFromProcessingData.R
\name{SaveInputFromProcessingData}
\alias{SaveInputFromProcessingData}
\title{Save Inputs from Processing Data}
\usage{
SaveInputFromProcessingData(inputData, edges)
}
\arguments{
\item{inputData}{A data.table containing the input data read from the
database, see ?getInputFromProcessData.  This is provided solely for
extraction of the appropriate column names, it isn't used in any other
way.}

\item{edges}{The edge data.table, after having been processed by the
AUPUS proceedure.  This is the updated data.set to write back to the
SWS.}
}
\value{
No data is returned, but the edges data.set is written back to
the SWS.
}
\description{
Save Inputs from Processing Data
}

