% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SavePopulationData.R
\name{SavePopulationData}
\alias{SavePopulationData}
\title{Save Population Data}
\usage{
SavePopulationData(populationData)
}
\arguments{
\item{populationData}{The population data to write back to the SWS.}
}
\value{
No data is returned, but the population data.set is written back
to the SWS.
}
\description{
Save Population Data
}

