% Generated by roxygen2 (4.0.1): do not edit by hand
\name{standardization}
\alias{standardization}
\title{This function performs standardization}
\usage{
standardization(graph, standardizeElement, aupusParam, plot)
}
\arguments{
\item{graph}{The graph object}

\item{standardizeElement}{The node attribute in the graph to be
standardized.}

\item{plot}{Whether the network should be plotted.}
}
\description{
Standardization is simply a weighted aggregation, but with
multiple hierachical levels.
}

