% Generated by roxygen2 (4.0.1): do not edit by hand
\name{standardizeNode}
\alias{standardizeNode}
\title{Function to standardize all the nodes in the node list of a graph}
\usage{
standardizeNode(graph, workingNode, standardizeAttributes)
}
\arguments{
\item{graph}{The graph object created by the function
constructGraph.}

\item{workingNodes}{The nodes to be standardized}

\item{standardizeAttributes}{The attribute of the nodes to be
standardized.}
}
\description{
The formula is hard coded
}

