% Generated by roxygen2 (4.0.1): do not edit by hand
\name{suaToNetworkRepresentation}
\alias{suaToNetworkRepresentation}
\title{This is the function to convert sua data to network specification
which is more natural for processing.}
\usage{
suaToNetworkRepresentation(extractionRateData, shareData, inputData, ratioData,
  balanceElementData, itemInfoData, populationData, aupusParam)
}
\arguments{
\item{extractionRateData}{The extraction rate data}

\item{shareData}{The share data from getShare function}

\item{inputData}{The input data from the function getInputFromProcessing.}

\item{ratioData}{The ratio data from getRatio}

\item{balanceElementData}{The balance element data from the
function getBalanceElementData.}

\item{itemInfoData}{}
}
\description{
This is the function to convert sua data to network specification
which is more natural for processing.
}

