% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subsetAupus.R
\name{subsetAupus}
\alias{subsetAupus}
\title{Subset AUPUS Data}
\usage{
subsetAupus(aupusData, itemKeys, aupusParam)
}
\arguments{
\item{aupusData}{The AUPUS dataset, typically as generated by
getAupusDataset.  This contains a list of 8 data.tables with the data of
interest.}

\item{itemKeys}{A numeric vector containing the item code(s) to filter on.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
An object of the same structure as aupusData, but with each
individual data.table filtered according to the set of keys provided.
}
\description{
This function takes as input the list of data.table data as generated by
getAupusDataset and subsets it based on specific item (commodity) codes.
This can be useful for analyzing smaller datasets and understanding what
the algorithm is doing.  Note that if the particular item code(s) have
any children or parents, those values (and their relatives, and so on)
will also be included.
}

