% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/symbolDescription.R
\docType{data}
\name{symbolDescription}
\alias{symbolDescription}
\title{Symbol Description}
\format{A data.table containing three columns, as described above.}
\usage{
data(symbolDescription)
}
\description{
A table of the symbols in use, containing the symbol/flag and the
name/description of the flag.  The replacable column indicates whether
data with that particular symbol can be imputed.
}
\keyword{flag,}
\keyword{replaceable}
\keyword{symbol,}

