% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/synchronizeByProducts.R
\name{synchronizeByProducts}
\alias{synchronizeByProducts}
\title{Synchronize By-Products}
\usage{
synchronizeByProducts(data, tree, standParams, specificTree = FALSE)
}
\arguments{
\item{data}{The data.table containing the full dataset for standardization.}

\item{tree}{The commodity tree which provides the edge structure.}

\item{standParams}{The parameters for standardization.  These parameters
provide information about the columns of data and tree, specifying (for
example) which columns should be standardized, which columns represent
parents/children, etc.}

\item{specificTree}{Logical.  If TRUE, it is assumed that there is a unique
  record in tree for each unique element of standParams$mergeKey.  If FALSE,
  extraction rates for each parent/child commodity are assumed to apply
  globally over all time.}
}
\value{
A data.table of the same form as data, but after the processing has
  been performed.
}
\description{
Synchronize By-Products
}

