% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transferSymb}
\alias{transferSymb}
\title{This function converts symbol when the it is transfer to a new
cell}
\usage{
transferSymb(symb)
}
\arguments{
\item{symb}{The original symbol}
}
\description{
This function converts symbol when the it is transfer to a new
cell
}

