% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/treeToStandardization.R
\name{treeToStandardization}
\alias{treeToStandardization}
\title{Tree to Standardization}
\usage{
treeToStandardization(aupusNetwork, aupusER, defaultOnly = FALSE)
}
\arguments{
\item{aupusNetwork}{The network object, consisting of a list of two objects:
nodes and edges.  Each individual object is a data.table object.  See
?suaToNetworkRepresentation.}

\item{aupusER}{The extraction rate table from the AUPUS data object.  This
argument is needed because it provides a way to overwrite the default
extraction rates with country specific ones, if available.}

\item{defaultOnly}{Logical.  Should only default extraction rates be used?
  Or should country-specific rates be used when available and default rates
  used in all other cases?  The old methodology isn't clear about what is
  used.}
}
\value{
The aupusNetwork, updated with the new edges and nodes.
}
\description{
During the AUPUS procedure, two classes of trees are used: the commodity
trees which specify how elements are to be processed down, and the
standardization trees which specify how commodities roll up.  This function
updates an aupus network (list of edges and ndoes) with a standardization
tree in place of the AUPUS tree.
}

