% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trendOnce}
\alias{trendOnce}
\title{This function trends the data once}
\usage{
trendOnce(Num, Symb, applyIndex = 1:length(Num), transfer = FALSE)
}
\arguments{
\item{Num}{The value column of the element}

\item{Symb}{The symbol column of the element}

\item{applyIndex}{Which value should be trended}

\item{transfer}{Should the symbol be converted by the function
transferSymb.}
}
\description{
This function trends the data once
}

