% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateEdges}
\alias{updateEdges}
\title{Function to update the extraction rates and input from processing
after each iteration of Aupus.}
\usage{
updateEdges(nodes, edges, element41Num, element131Num, aupusParam)
}
\arguments{
\item{nodes}{The nodes data returned by the function buildNodes}

\item{edges}{The edge data returned by the function buildEdges}

\item{element41Num}{The column correspoinding to element 41, or
extraction rates.}

\item{element131Num}{The column corresponding to element 131.}
}
\description{
Function to update the extraction rates and input from processing
after each iteration of Aupus.
}

