% Generated by roxygen2 (4.0.1): do not edit by hand
\name{updateInputFromProcessing}
\alias{updateInputFromProcessing}
\title{The function re-calculates the input from processing}
\usage{
updateInputFromProcessing(nodes, edges, element31Num, aupusParam)
}
\arguments{
\item{nodes}{The nodes data returned by the function buildNodes}

\item{edges}{The edge data returned by the function buildEdges}

\item{element131Num}{The column corresponding to element 131.}
}
\description{
The function re-calculates the input from processing
}

