% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/US.R
\docType{data}
\name{US}
\alias{US}
\title{US Sample dataset}
\format{A list object containing 8 individual datasets.  Each dataset is
generated by the corresponding get* function in the faoswsAupus package.
Most elements are data.tables, but some (such as the ratio and
balanceElement datasets) are lists of data.tables.  See the documentation
for the corresponding get* functions for more details.}
\usage{
data(US)
}
\description{
An example dataset of what can be created from calling the function
getAupusDataset.
}
\keyword{aupus,}
\keyword{dataset}

