% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wildCardFill}
\alias{wildCardFill}
\title{The function takes wild card data to fill in original data for a
particular variable.}
\usage{
wildCardFill(originalData, wildCardData, variable, verbose = FALSE, ...)
}
\arguments{
\item{originalData}{The data to be filled in}

\item{wildCardData}{The wild card data, see ratio or shares data
extracted from the getShare and getRatio function.}

\item{variable}{The name of the variable to be filled.}
}
\description{
Shares, ratio and balance element have area and year specific
rates, but at the same time they have wild card values which are
to be applied when the specific rates are not available.
}
\details{
This function fills in the gap with wild card values when year
specific values are not available.
}

