% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/appendSymbol.R
\name{appendSymbol}
\alias{appendSymbol}
\title{Function to fill symbol based on calculated value}
\usage{
appendSymbol(value, newSymbol, missingSymbol = "M")
}
\arguments{
\item{value}{The vector of calculated values}

\item{newSymbol}{The new symbol to be assigned}

\item{missingSymbol}{The symbol filled in when value is missing}
}
\value{
A list of two elements: the vector of passed values and the new
derived vector of symbols.
}
\description{
The function will take a vector of values and fill in the new
symbol if the value is not missing. If the value is missing, it
will insert the missing symbol.
}

