% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Aupus.R
\name{Aupus}
\alias{Aupus}
\title{AUPUS}
\usage{
Aupus(aupusNetwork, aupusParam)
}
\arguments{
\item{aupusNetwork}{A list of two data.tables.  The objects should be the
node and edge data.tables.  Typically, this argument will be the output
from the function suaToNetworkRepresentation.}

\item{aupusParam}{A list of running parameters to be used in pulling the
data. Typically, this is generated from getAupusParameter (see that
function for a description of the required elements).}
}
\value{
A list with two components: nodes and edges.  These objects are the
same as the input nodes and edges but after being processed by the AUPUS
algorithm.
}
\description{
This function performs the whole aupus procedure.
}

