% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildEdges.R
\name{buildEdges}
\alias{buildEdges}
\title{Build Edges}
\usage{
buildEdges(dataList)
}
\arguments{
\item{dataList}{A list of AUPUS datasets to be analyzed, typically as
produced by the function getAupusDataset.  For more details on this
argument, please see ?getAupusDataset.}
}
\value{
A data.table object with the same keys as dataList$inputData
(typically geographicAreaFS, measuredItemParentFS, measuredItmChildFS, and
timePointYearsSP).  Additional columns are Value_share (shares as specified
in dataList$shareData), Value_extraction (extraction rates as specified in
dataList$extractionRateData), and Value_input, flagFaostat_input (as
specified in dataList$inputData).
}
\description{
Function to merge shares and extraction rate and build the
relation graph.
}

