% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildNodes.R
\name{buildNodes}
\alias{buildNodes}
\title{Build Nodes}
\usage{
buildNodes(dataList)
}
\arguments{
\item{dataList}{A list of AUPUS datasets to be analyzed, typically as
produced by the function getAupusDataset.  For more details on this
argument, please see ?getAupusDataset.}
}
\value{
A data.table with the node data.  Essentially, this is a data.table
containing all the variables meant to be stored in the nodes: AUPUS, input,
ratio, balance element, item info, and population.
}
\description{
This function creates the node structure by merging aupus, ratio, balance
element, item info and population datasets.
}

