% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateAupusElements.R
\name{calculateAupusElements}
\alias{calculateAupusElements}
\title{Calculate AUPUS Elements}
\usage{
calculateAupusElements(aupusFinalData, itemTypeCol, balanceElementNum,
  aupusParam)
}
\arguments{
\item{aupusFinalData}{The data.table object containing the nodes data,
typically as created by the suaToNetworkRepresentation function.}

\item{itemTypeCol}{The column name of aupusFinalData corresponding to the
item type of the commodity item.}

\item{balanceElementNum}{The column name of aupusFinalData corresponding to
the balance element column.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
No value is returned.  Instead, the passed data.table
(aupusFinalData) has each of it's elements computed and is balanced.
}
\description{
This is the wrapper function that performs the full AUPUS module.  It
computes each individual element, then proceeds to compute utilization and
balances.  This function should be called several times to process all the
processing levels of the items.
}

