% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateBalance.R
\name{calculateBalance}
\alias{calculateBalance}
\title{Calculate Balance}
\usage{
calculateBalance(supply, utilization, balanceElement, data, aupusParam)
}
\arguments{
\item{supply}{The column name of data corresponding to the calculated total
supply.}

\item{utilization}{The column correspoding to calculated total utilization.}

\item{balanceElement}{The column name of data corresponding to the
balancing element.}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
A logical vector indicating which rows were updated with the
balance.
}
\description{
The function will calculate the balance (in most cases, the difference
between supply and utilization) and fill this value into the balanceElement
column.  However, if this computed value is negative, or if the
balanceElement symbol is not replaceable (see ?replaceable) then this
computed value is placed in the statistical discrepancy column.
}

