% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle101.R
\name{calculateEle101}
\alias{calculateEle101}
\title{Calculate Element 101 (Use for Animals)}
\usage{
calculateEle101(stotal, data, aupusParam)
}
\arguments{
\item{stotal}{The column name of data corresponding to the total supply.}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 101's value and
symbol updated.
}
\description{
The amount used for animal feed is computed using a very simple formula:
(Element 101) = (Ratio 101) * (Total Supply) / 100.
Essentially, the amount used for animals is computed as a simple ratio of
the total supply.  This function performs that calculation.
}

