% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle11.R
\name{calculateEle11}
\alias{calculateEle11}
\title{Calculate Element 11 (Initial Existence)}
\usage{
calculateEle11(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 11's value and
symbol updated.
}
\description{
Initial existence (element 11) at time t is imputed with final existence
(element 161) at time t-1, but imputation only occurs if data is missing.
Note that not all item types are not updated in this way (see code).
}

