% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle111.R
\name{calculateEle111}
\alias{calculateEle111}
\title{Calculate Element 111 (Use for Same Product)}
\usage{
calculateEle111(stotal, data, aupusParam)
}
\arguments{
\item{stotal}{The column name of data corresponding to the total supply.}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns a list (of length 2) of integer vectors
containing the row numbers of observations which were updated (the first
vector when a value was computed using ratio 111, the second for when
ratio 171 was used).  However, it also has a side effect: rows within the
passed data.table ("data") have element 111's value and symbol updated.
}
\description{
This function imputes element 111, which can be thought of as the amount
used for seeding.  The rules are as follows:
}
\details{
\itemize{
    \item If Ratio 171 (presumably a seeding ratio) is not available for
    the current commodity, a similar formula to 101 is used:
    E_111 = R_111 * (total supply) / 100.
    \item If Ratio 171 is available, then an estimate for total production
    is used to multiply by Ratio 171 to compute the amount used for seed.
    The preferred estimate for production value is element 21 (potential
    producing factor) in the next year; if this is unavailable then element
    31 (actual producing factor) in the next year is used; then element 21
    in the current year; and lastly element 31 in the current year.  The
    value for element 111 is then filled in as this computed seed value
    times R_171 divided by 1000.
}
}

