% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle121.R
\name{calculateEle121}
\alias{calculateEle121}
\title{Calculate Element 121 (Losses)}
\usage{
calculateEle121(stotal, data, aupusParam)
}
\arguments{
\item{stotal}{The column name of data corresponding to total supply}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector
containing the row numbers of observations which were updated.
However, it also has a side effect: rows within the
passed data.table ("data") have element 121's value and symbol updated.
}
\description{
Element 121 (losses) is calculated as a simple ratio of total supply:
E_121 = R_121 * (total supply) / 100.
Values are always updated in this fashion unless the symbol for element 121
is not replaceable (i.e. it's an official figure).
}

