% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle141.R
\name{calculateEle141}
\alias{calculateEle141}
\title{Calculate Element 141 (Consumption)}
\usage{
calculateEle141(stotal, data, aupusParam)
}
\arguments{
\item{stotal}{The column name of data corresponding to total supply.}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector
containing the row numbers of observations which were updated.
However, it also has a side effect: rows within the
passed data.table ("data") have element 141's value and symbol updated.
}
\description{
This function computes element 141 (Consumption)
}
\details{
The following AUPUS logic is used in computing this element:

\itemize{
    \item For commodity types in 58, 59, 60, and 61, element 141 is
    computed as E_11 + E_51 + E_61 - E_91 - E_95 - E_161.
    \item For commodity type 50 (Jute), the original AUPUS logic was very
    complex and also written in such a way as to cause potential errors.
    Thus, no processing is done for this commodity type: it is left
    unchanged by this function.
    \item For all other commodity types, the consumption is computed as a
    ratio of the total supply: E_141 = R_141 * (total supply) / 100
}
}

