% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle151.R
\name{calculateEle151}
\alias{calculateEle151}
\title{Calculate Element 151 (Reemployment other sector)}
\usage{
calculateEle151(stotal, data, aupusParam)
}
\arguments{
\item{stotal}{The column name of data corresponding to total supply.}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector
containing the row numbers of observations which were updated.
However, it also has a side effect: rows within the
passed data.table ("data") have element 151's value and symbol updated.
}
\description{
In almost all cases, element 151 (Reemployment other sector) is computed as
a ratio of the total supply: E_151 = R_151 * (total supply) / 100.
However, for one particular commodity (code 1687, Charcoal), element 151 is
instead computed as the value of element 131 for commodity 1648 minus the
value of element 51 for commodity 1687.
}

