% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle161.R
\name{calculateEle161}
\alias{calculateEle161}
\title{Calculate Element 161 (Final existence)}
\usage{
calculateEle161(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 161's value and
symbol updated.
}
\description{
This element is only computed for commodity type 57 (sugar).  This element
is calculated as element 11 (Initial Existence) + element 71 (From Initial
Existence).
}

