% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle171.R
\name{calculateEle171}
\alias{calculateEle171}
\title{Calculate Element 171 (Consumption, Sugar only)}
\usage{
calculateEle171(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 171's value and
symbol updated.
}
\description{
This element is only calculated for sugar.  Moreover, it's just a summation
of 5 other commodities: 101, 121, 131, 141, and 151.
}

