% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle174.R
\name{calculateEle174}
\alias{calculateEle174}
\title{Calculate Element 174 (Consumption per Caput, Sugar only)}
\usage{
calculateEle174(population11Num, data, aupusParam)
}
\arguments{
\item{population11Num}{The column corresponds to element 11 of the
population.}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 174's value and
symbol updated.
}
\description{
This function computes consumption per caput for sugar.  It is simply the
ratio of total consumption over total population, and so the calculation
is element 171 divided by population 11.
}

