% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle21.R
\name{calculateEle21}
\alias{calculateEle21}
\title{Calculate Element 21 (Potential Producing Factor)}
\usage{
calculateEle21(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector
containing the row numbers of observations which were updated.  However, it
also has a side effect: rows within the passed data.table ("data") have
element 21's value and symbol updated.
}
\description{
The rule for replacement here is
(element 21) = (element 111) * 1000 / (ratio 171)
}

