% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle31.R
\name{calculateEle31}
\alias{calculateEle31}
\title{Calculate Element 31}
\usage{
calculateEle31(inputNum, data, aupusParam)
}
\arguments{
\item{inputNum}{The column name of data corresponding to inputs from
processing.}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns a list (of length two) of integer vectors
containing the row numbers of observations which were updated (first by
computing as a function of inputs, then next by setting to NA).  However,
it also has a side effect: rows within the passed data.table ("data") have
element 31's value and symbol updated.
}
\description{
The value of this element is set to the sum of all the input values
from the inputFromProcessing table (see getInputFromProcessData).
Otherwise, the value (if replaceable) is set to NA.
}

