% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle41.R
\name{calculateEle41}
\alias{calculateEle41}
\title{Calculate Element 41 (Productivity Element)}
\usage{
calculateEle41(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 41's value and
symbol updated.
}
\description{
Calculation of element 41 is done by simply multiplying ratio 41 by 100.
Values which are replaceable are updated by this function.
}

