% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle51.R
\name{calculateEle51}
\alias{calculateEle51}
\title{Calculate Element 51}
\usage{
calculateEle51(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 51's value and
symbol updated.
}
\description{
There is no calculation of element 51 (Output) for almost all cases.  Two
special cases exist, however:
\itemize{
    \item If the commodity item type is in aupusGroups$escrSugarProduction
    and if element 58 (Production Crop Year) exists for the commodity of
    interest, then element 51 is set to the value of element 58.
    \item If the commodity code is 3183, then element 51 (Output) is set
    to the sum of element 51 (Output) for commodity codes 3158 and 3159.
}
}

