% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle541.R
\name{calculateEle541}
\alias{calculateEle541}
\title{Calculate Element 541 (Final Demand)}
\usage{
calculateEle541(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector
containing the row numbers of observations which were updated.
However, it also has a side effect: rows within the passed
data.table ("data") have element 541's value and symbol updated.
Additionally, columns will be added to the passed data.table if they don't
currently exist: Value_measuredElementFS_54X, where X = 2, 3, 4, 5.
}
\description{
Element 541 (Final Demand) is computed as a summation of elements 542, 543,
544, and 545.  If some of those values are missing, they are ignored in
the calculation.  However, if all four of the values are missing, then the
resulting calculation returns a 0M for element 541.
}

