% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle546.R
\name{calculateEle546}
\alias{calculateEle546}
\title{Calculate Element 546 (Total Demand)}
\usage{
calculateEle546(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector
containing the row numbers of observations which were updated.
However, it also has a side effect: rows within the passed
data.table ("data") have element 546's value and symbol updated.
}
\description{
Element 546 (Total Demand) is computed as a summation of elements 541, 151,
and 91  If some of those values are missing, they are ignored in
the calculation.  However, if all three of the values are missing, then the
resulting calculation returns a 0M for element 546.
}

