% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle58.R
\name{calculateEle58}
\alias{calculateEle58}
\title{Calculate Element 58 (Production Crop Year)}
\usage{
calculateEle58(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns an integer vector containing the row numbers
of observations which were updated.  However, it also has a side effect:
rows within the passed data.table ("data") have element 58's value and
symbol updated.
}
\description{
Calculation of element 58 is done by adding together element 3158 and 3159
for the source commodities.
}

