% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle63.R
\name{calculateEle63}
\alias{calculateEle63}
\title{Calculate Element 63}
\usage{
calculateEle63(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns a list (of length 2) of integer vectors
containing the row numbers of observations which were updated (the first
vector when a value was computed, the second for when 62 or 61 was
missing).  However, it also has a side effect: rows within the passed
data.table ("data") have element 63's value and symbol updated.
}
\description{
This function calculates element 63 (Inflow Unit Value) using element 61
(Inflow) and element 62 (Inflow Value).  The calculation is simple: unit
value (63) should be the total value (62) divided by the total quantity
(61).  However, a multiplicative constant of 1000 is also used, so the
formula is 63 = 62 * 1000 / 61.
}
\details{
Note that if 61 or 62 are not available, and if 63 is replaceable, then it
is replaced with a missing value (0M).
}

