% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle66.R
\name{calculateEle66}
\alias{calculateEle66}
\title{Calculate Element 66 (standardized inflow)}
\usage{
calculateEle66(shares, data, shareData)
}
\arguments{
\item{shares}{The column name of the share variable in shareData (i.e. the
column containing the proportion of the parent commodity that is passed to
the child commodity).}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{shareData}{The data.table containing the shares data.}
}
\value{
This function is not currently working.
}
\description{
This function does not seem to be run.  It contains functions which are not
defined, and the documentation states that computation of this element is
only for commodities that are generally not used within AUPUS, and that are
only trade commodity types.
}

