% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateEle93.R
\name{calculateEle93}
\alias{calculateEle93}
\title{Calculate Element 93}
\usage{
calculateEle93(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function returns a list (of length 2) of integer vectors
containing the row numbers of observations which were updated (the first
vector when a value was computed, the second for when 62 or 61 was
missing).  However, it also has a side effect: rows within the passed
data.table ("data") have element 63's value and symbol updated.
}
\description{
This function calculates element 93 (outflow unit value).  The logic is
very similar to the calculation for element 63 (inflow unit value) in that
unit value (93) should be the total value (92) divided by the total
quantity (91).  Also, the multiplicative constant of 1000 is still used in
this case, so the formula is 93 = 92 * 1000 / 91.
}
\details{
However, there is one difference between computation of element 93 and
element 63: computation is not done for elements 42 through 52.
}

