% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateTotalInput.R
\name{calculateTotalInput}
\alias{calculateTotalInput}
\title{Calculate Total Input}
\usage{
calculateTotalInput(inputData, inputNum, aupusParam)
}
\arguments{
\item{inputData}{The data containing the input extracted from the
function getInputFromProcess after running the
updateInputFromProcess function.}

\item{inputNum}{???}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
This function is not currently used.
}
\description{
It's not clear what this function was intended for, as it is not
called by any other functions in this package.
}

