% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateTotalUtilization.R
\name{calculateTotalUtilization}
\alias{calculateTotalUtilization}
\title{Calculate Total Utilization}
\usage{
calculateTotalUtilization(data, aupusParam)
}
\arguments{
\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
No value is returned, but a new column ("TOTAL_UTILIZATION") is
appended to the passed data.table.
}
\description{
This function calculates total utilization for each row by summing up all
elements which correspond to utilization.  These are currently elements
91 (Outflow), 96 (Standardized Outflow), 101 (Use for Animals), 111 (Use
for Same Product), 121 (Losses), 131 (Reemployment), 141 (Consumption),
151 (Reemployment Other Sector), 161 (Final Existence), and 546 (Total
Demand).
}
\details{
Also, element 151 is not considered utilization for item types defined by
aupusGroups$reemploymentNotAsUtilization (currently only item type 53).
}

