% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkShareUnity.R
\name{checkShareUnity}
\alias{checkShareUnity}
\title{This is a function checking whether the shares sums up to 100.}
\usage{
checkShareUnity(shareData)
}
\arguments{
\item{shareData}{The share data returned from getShare.}
}
\value{
This function is not currently used.
}
\description{
This is a function checking whether the shares sums up to 100.
}

