% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collapseShare.R
\name{collapseShare}
\alias{collapseShare}
\title{Collapse Share}
\usage{
collapseShare(aupusParam, shareData, shares, verbose = FALSE)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{shareData}{The shares data obtained from the function getShare.  This
is a list of length 3 with data.tables as elements (specific, yearWildCard,
and areaYearWildCard).  Each data.table contains share information at a
different level of specificity.}

\item{shares}{The name of the column correspond to shares within each
element of shareData (note that it must be the same for each element).}

\item{verbose}{Whether the output should be printed.}
}
\value{
Returns a single data.table derived by condensing the shareData
list into one dataset.
}
\description{
Share data is extracted from the database in three chunks: country/year
specific data, country specific data, and generic data (applying to all
countries/years).  The most specific data available should be used when
available, otherwise a more general value should be used.  This function
collapses the three different datasets (supplied in shareData) into one
final dataset.
}

