% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constructStandardizationGraph.R
\name{constructStandardizationGraph}
\alias{constructStandardizationGraph}
\title{Construct Standardization Graph}
\usage{
constructStandardizationGraph(aupusNetwork, standardizeElement, aupusParam)
}
\arguments{
\item{aupusNetwork}{A list of 2 data.tables named "nodes" and "edges".
These are usually produced by suaToNetworkRepresentation and contain
information about the data at the nodes and along the edges of the network.}

\item{standardizeElement}{The column names of nodes that should be
included in the returned graphs.  These are the elements to be
standardized.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
A list of the same length as the number of unique years in the
aupus network.
}
\description{
This function constructs the graph/network required for standardization by
generating one graph object for each year.  This graph has the supplied
edges and nodes as the corresponding edges and nodes on the graph.
}

