% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faoswsAupus-package.R
\docType{package}
\name{faoswsAupus-package}
\alias{faoswsAupus-package}
\title{Package to perform the AUPUS calculation.}
\description{
This package provides functions to replicate the old system methodology
(AUPUS) in the new system.
}
\author{
Joshua M. Browning \email{joshua.browning@fao.org}, Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

