% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fbsStandardization.R
\name{fbsStandardization}
\alias{fbsStandardization}
\title{FBS Standardization}
\usage{
fbsStandardization(graph, standardizeElement, plot, aupusParam)
}
\arguments{
\item{graph}{The standardization graph from the function
constructStandardizationGraph.}

\item{standardizeElement}{A character vector of the column names of the
node object that should be processed.  This should correspond to the
elements of the FBS that need to be standardized.}

\item{plot}{Whether the network should be plotted as it is processed.
A plot will be generated for each year/processing level, with prompts
between, so only use this option if running this function in interactive
mode.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
A data.table object containing the final standardization values
for each year.
}
\description{
This function takes the graph and performs the standardization.
}

