% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fillBalance.R
\name{fillBalance}
\alias{fillBalance}
\title{Fill Balance}
\usage{
fillBalance(calculatedBalance, balanceNum, balanceSymb, element181Num,
  element181Symb, data)
}
\arguments{
\item{calculatedBalance}{The column name of data corresponding to the
balance calculated by the function calculateBalance.}

\item{balanceNum}{The column name of data corresponding to the value of
the balancing element.}

\item{balanceSymb}{The column name of data corresponding to the symbol
column for the balancing element.}

\item{element181Num}{The column name of data corresponding to the value of
element 181 (Statistical Discrepancies).}

\item{element181Symb}{The column name of data corresponding to the symbol
of element 181 (Statistical Discrepancies).}

\item{data}{The AUPUS node dataset, typically as produced by buildNodes.}
}
\value{
A list with 5 vectors of the same length:
\itemize{
    \item originalValue: The numeric vector of the updated balance column.
    \item originalSymb: The symbol vector of the updated balance column.
    \item discrepancyValue: The numeric vector for the statistical
    discrepancy column.
    \item discrepancySymb: The symbol vector for the statistical
    discrepancy column.
    \item replaced: A logical vector which is TRUE if the computed value
    has been placed in the balance element column and FALSE otherwise.
}
}
\description{
This function determines how the balance and statistical discrepancy
columns should be filled in.  There are two scenarios:
}
\details{
\itemize{
    \item If the symbol currently in the balance column is "replaceable"
    (see ?replaceable) and if the computed balance is positive, then the
    computed balance is placed in the balance column and statistical
    discrepancy is set to 0.
    \item If the computed balance is negative, or if the current balance
    symbol is not replaceable, then the computed balance is placed in the
    statistical discrepancy column.
}
}

