% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fillMissingColumn.R
\name{fillMissingColumn}
\alias{fillMissingColumn}
\title{This function creates columns which are missing.}
\usage{
fillMissingColumn(data, allColumn)
}
\arguments{
\item{data}{The data}

\item{allColumn}{The theoretical set of all column names.}
}
\value{
This function does not seem to be currently in use.
}
\description{
Some countries may not have a certain element calculated and thus
when the data is extracted, the column is missing. This function
will recreate the missing column.
}

