% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findConnectedGraph.R
\name{findConnectedGraph}
\alias{findConnectedGraph}
\title{Find Connected Graph}
\usage{
findConnectedGraph(graph, commodity)
}
\arguments{
\item{graph}{The graph object created by the function
constrctGraph.}

\item{commodity}{The commodities}
}
\value{
This function is not currently used.
}
\description{
This function takes a graph/network and then find the subset of
the network which are connected to the specified nodes/commodities.
}

