% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAupusData.R
\name{getAupusData}
\alias{getAupusData}
\title{Get AUPUS Data}
\usage{
getAupusData(aupusParam, database = c("new", "old"), conn)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{database}{Whether to use the new or the old statistical
working system.}

\item{conn}{The RJDBS connection to the old working system, only required
if database = "old".}
}
\value{
A data.table object containing the AUPUS data.  The first three
columns contain the dimensions (geographicAreaFS, measuredItemFS, and
timePointYears) and the remaining columns are values and flags for each of
the elements required by AUPUS.
}
\description{
This function extracts the AUPUS data from the data base.
}

