% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAupusDataset.R
\name{getAupusDataset}
\alias{getAupusDataset}
\title{Get AUPUS Data}
\usage{
getAupusDataset(aupusParam, assignGlobal = FALSE)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{assignGlobal}{Whether the data should be assigned globally (TRUE)
or returned as a list (FALSE). Defaults to FALSE.}
}
\value{
If assignGlobal = FALSE, then a list is returned with names
"aupusData", "inputData", "ratioData", "shareData", "balanceElementData",
"itemInfoData", "populationData", and "extractionRateData".  Each of these
elements is a data.table containing information for the AUPUS processing.
If assignGlobal = TRUE, then these data.tables are assigned to the global
environment and no data is returned.
}
\description{
This function obtains all required data for the AUPUS module by running
several individual functions, such as getRatioData, getShareData, etc.  The
datasets returned by these individual functions are then returned from this
function in a list (if assignGlobal = FALSE) or are assigned to the global
environment (if assignGlobal = TRUE).
}

