% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getAupusParameter.R
\name{getAupusParameter}
\alias{getAupusParameter}
\title{Get AUPUS Parameters}
\usage{
getAupusParameter(areaCode, assignGlobal = TRUE, yearsToUse = NULL)
}
\arguments{
\item{areaCode}{A character value giving the country code of the country of
interest.}

\item{assignGlobal}{logical, default to FALSE, in which case a list is
returned. If TRUE, then the result will be assigned globally.}

\item{yearsToUse}{If NULL, then all available years (as determined by
GetCodeList) will be used.  Otherwise, a vector of years should be provided
and these years alone will be included.}
}
\value{
If assignGlobal is TRUE, nothing is returned but 5 objects are
written to the global environment: areaCode, itemCode, elementCode, year,
and keyNames.  The first four of these variables each define a dimension
for slicing the data, and the fifth element is a list of arguments to
processing functions.  If assignGlobal is false, these 5 objects are
returned in a named list.
}
\description{
This function gets all the parameters in order to query the data, and
either assigns these parameters to the global environment of returns them
as a list.
}

