% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getBalanceElementData.R
\name{getBalanceElementData}
\alias{getBalanceElementData}
\title{Get Balance Element Data}
\usage{
getBalanceElementData(aupusParam, database = c("new", "old"), conn)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{database}{Whether to use the new or the old statistical
working system.}

\item{conn}{The RJDBS connection to the old working system.  Only required
if database = "old".}
}
\value{
A list of three items: specific, yearWildCard, and
areaYearWildCard.  Each of these elements has the same general structure:
a data.table with zero to two dimension columns (area and year), the
measured item, and then the element that item is balanced into.
}
\description{
This function extracts the balance element data from the database.  This
data specifies, for each measured item, the final element that it will be
balanced to.  For example, Flour of Wheat (item = 16) is placed into Food
(element = 141).
}

