% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getInputFromProcessData.R
\name{getInputFromProcessData}
\alias{getInputFromProcessData}
\title{Get Input from Processing Data}
\usage{
getInputFromProcessData(aupusParam, database = c("new", "old"), conn)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{database}{Whether to use the new or the old statistical
working system.}

\item{conn}{The RJDBS connection to the old working system.  Only required
if database = "old".}
}
\value{
This function returns a data.table object containing the input from
processing data.
}
\description{
Some data used in AUPUS will be specified by the users prior to the AUPUS
routine (for example, official data provided by some countries).  This
function provides a way for this input data to be specified and utilized
in AUPUS.
}

