% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPopulationData.R
\name{getPopulationData}
\alias{getPopulationData}
\title{Get Population Data}
\usage{
getPopulationData(aupusParam, database = c("new", "old"), conn)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{database}{Whether to use the new or the old statistical
working system.}

\item{conn}{The RJDBS connection to the old working system.  Only required
if database = "old".}
}
\value{
A data.table containing the country and year values specified in
aupusParam and value and flags for the population (elements 11 and 21).
}
\description{
This is the function to obtain the population separately.
}

