% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getRatioData.R
\name{getRatioData}
\alias{getRatioData}
\title{Get Ratio Data}
\usage{
getRatioData(aupusParam, database = c("new", "old"), conn)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{database}{Whether to use the new or the old statistical
working system.}

\item{conn}{The RJDBS connection to the old working system.  Only required
if database = "old".}
}
\value{
A list of three items: specific, yearWildCard, and
areaYearWildCard.  Each of these elements has the same general structure:
a data.table with one to three dimension columns (area, year, and item) and
then many columns containing the ratios (or conversion factor) from that
item into elements in the FBS.  If such values don't make sense or don't
apply, an NA is in the data.
}
\description{
This function extracts the ratio data from the data base.  The ratio data
contains information on ratios that exist between certain "elements" and
"items."  For example, how many calories (element = "261") are in wheat
flour (item = "16")?  The answer according to the database is 364, but
better estimates may exist for specific years and specific countries.  This
dataset contains the overall average rates and the specific rates.
}

