% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getShareData.R
\name{getShareData}
\alias{getShareData}
\title{Get Share Data}
\usage{
getShareData(aupusParam, database = c("new", "old"), conn)
}
\arguments{
\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}

\item{database}{Whether to use the new or the old statistical
working system.}

\item{conn}{The RJDBS connection to the old working system.  Only required
if database = "old".}
}
\value{
A list of three items: specific, yearWildCard, and
areaYearWildCard.  Each of these elements has the same general structure:
a data.table with zero to two dimension columns (area and year), the parent
and child items, and then the Value_share column (representing the
proportion allocated to each child).
}
\description{
This function extracts the shares data from the data base.  The shares data
contains information on children are shared amongst parents.  For example,
some children of wheat flour (item = "16") are: macaroni (18), bread (20),
and pastry (22).  In Algeria (geographicAreaFS = "4"), the shares are 95,
3, and 2 which means that 95% of wheat flour is converted to macaroni, 3%
to bread and 2% to pastry.  A default share is also available with the 0
wildcard, and values can vary by year (also with a wildcard of 0).
}

