% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numberOfMissingElement.R
\name{numberOfMissingElement}
\alias{numberOfMissingElement}
\title{Number of Missing Elements}
\usage{
numberOfMissingElement(...)
}
\arguments{
\item{...}{The objects}
}
\value{
A vector of the number of total inputs missing for each ith
element.
}
\description{
This function takes several inputs, typically vectors, and returns a vector
whose ith element specifies the number of input vectors with missing values
in the ith position.
}
\examples{
numberOfMissingElement(c(1,2,NA,NA,5),
                       c(1,NA,2,3,NA),
                       c(1,2,NA,4,NA))
}

