% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numberofTrendingElement.R
\name{numberOfTrendingElement}
\alias{numberOfTrendingElement}
\title{Number of Trended Elements}
\usage{
numberOfTrendingElement(...)
}
\arguments{
\item{...}{The objects}
}
\value{
A vector of the number of trended symbols in the ith position.
}
\description{
This function takes several inputs, typically vectors, and returns a vector
whose ith element specifies the number of input vectors with symbol equal
to "T" in the ith position (which denotes a trended estimate).
}
\examples{
numberOfTrendingElement(c("M","","","T",""),
                        c("M", "T", "M", "T", ""),
                        c("C", "C", "T", "T", "T"))
}

