% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replaceable.R
\name{replaceable}
\alias{replaceable}
\title{Returns the index of which values can be replaced}
\usage{
replaceable(symb, newValue = 0, replaceableSymb = c("C", "T", "M"))
}
\arguments{
\item{symb}{The vector of symbols to be tested.}

\item{newValue}{The new value which can be written to the cell.  The AUPUS
logic states that the old value should not be replaced if it was trended
and the new value is missing (0 in the old framework).}

\item{replaceableSymb}{A vector specifying which symbols can be replaced.}
}
\value{
A logical vector indicating if each individual element can be
replaced.
}
\description{
Only data with replaceable symbols are replaced by the
module. This function will return the index in which the values
can be replaced.
}

