% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standardization.R
\name{standardization}
\alias{standardization}
\title{Standardization}
\usage{
standardization(graph, standardizeElement, plot, aupusParam, ...)
}
\arguments{
\item{graph}{The igraph object containing the node and edge information,
typically as generated by}

\item{standardizeElement}{A character vector of the column names of the
node object that should be processed.  This should correspond to the
elements of the FBS that need to be standardized.}

\item{plot}{Whether the network should be plotted as it is processed.
Several plots will be generated with prompts between, so only use this
option if running this function in interactive mode.}

\item{aupusParam}{A list of running parameters to be used in pulling the
data. Typically, this is generated from getAupusParameter (see that
function for a description of the required elements).}

\item{...}{Additional arguments to be passed to the plotting function.}
}
\value{
A data.table object with one column for the item key and one
column for each element of standardizeElement (containing the value
after standardization has been performed).
}
\description{
Standardization is simply a weighted aggregation, but with
multiple hierachical levels.  This function contains the logic for
iterating over the different hierarchical levels, but it passes the
actual computation of the aggregations on to the standardizeNode
function.
}

