% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standardizeNode.R
\name{standardizeNode}
\alias{standardizeNode}
\title{Standardize Node}
\usage{
standardizeNode(graph, workingNode, standardizeElement)
}
\arguments{
\item{graph}{The graph object created by the function
constructGraph.}

\item{workingNode}{The nodes to be standardized}

\item{standardizeElement}{The attribute of the nodes to be
standardized.}
}
\value{
A graph object that has been updated according to the description
above.
}
\description{
Standardization follows the below process:
\itemize{
    \item Compute a matrix ("reverseMatrix" in the code) describing how
    much of a particular quantity is passed from a child to it's parent.
    The (i,j)-th entry of this matrix specifies a conversion factor for
    transferring element j quantities into element i quantities.
    \item Compute the matrix of required elements (column) for each
    commodity (row), and call this "valueMatrix".
    \item Compute the new value for each commodity by pre-multiplying
    valueMatrix by reverseMatrix and therefore aggregating children
    commodities into their parents.
    \item Lastly, save this new data back to the passed graph object, and
    return.
}
}

