% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/suaToNetworkRepresentation.R
\name{suaToNetworkRepresentation}
\alias{suaToNetworkRepresentation}
\title{SUA to Network Representation}
\usage{
suaToNetworkRepresentation(dataList, aupusParam)
}
\arguments{
\item{dataList}{A list of AUPUS datasets to be analyzed, typically as
produced by the function getAupusDataset.  For more details on this
argument, please see ?getAupusDataset.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
A list of two elements: nodes and edges.  Both of these objects are
data.tables, and they share the same keys with one exception: edges has a
parent and child item key where nodes only has an item key.  All the input,
share and extraction data is stored in edges and the AUPUS, ratio, balance,
item, and population data is stored in the nodes.
}
\description{
SUA data (Supply Utilization Accounts) contains information on specific
uses of commodities within the food balance sheet.  Often, SUA is explained
as being analogous to all the financial transactions of a company and the
Food Balance Sheets are thought of as the financial summary statement.
}
\details{
This function converts sua data to a network specification
which is more natural for processing.
}

