% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transferSymb.R
\name{transferSymb}
\alias{transferSymb}
\title{Transfer Symbol}
\usage{
transferSymb(symb)
}
\arguments{
\item{symb}{The original symbol to be updated.}
}
\value{
The updated symbol.
}
\description{
This function contains the logic for updating a symbol when it is
transfered to a new cell.
}

