% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trendOnce.R
\name{trendOnce}
\alias{trendOnce}
\title{Trend Once}
\usage{
trendOnce(Num, Symb, applyIndex = 1:length(Num), transfer = FALSE)
}
\arguments{
\item{Num}{A vector of numeric values to be trended.}

\item{Symb}{A vector of the same length as Num containing the symbols (or
"flags") for the observations in Num.}

\item{applyIndex}{This argument allows the user to restrict which values
can be trended.  By default, all values are included.}

\item{transfer}{Should the symbol be converted by the function
transferSymb.}
}
\value{
A list containing the values and symbols after the trending has
occurred.
}
\description{
"Trending" the data refers to carrying over past values.  For example,
trending c(1,2,NA,3) would yield c(1,2,2,3).  However, missing values are
only replaced if their corresponding symbols are "T" or "C".
}
\details{
This function trends the passed data (Num) assuming a vector of symbols
(Symb).  The "transfer" argument allows the user to specify if the symbol
should also be transferred when the number is copied forward.
}

