% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/updateEdges.R
\name{updateEdges}
\alias{updateEdges}
\title{Update Edges}
\usage{
updateEdges(nodes, edges, aupusParam)
}
\arguments{
\item{nodes}{The nodes data returned by the function buildNodes}

\item{edges}{The edge data returned by the function buildEdges}

\item{aupusParam}{A list of running parameters to be used in pulling the
data. Typically, this is generated from getAupusParameter (see that
function for a description of the required elements).}
}
\value{
No data is returned, but the edges object is updated with the
information from nodes.
}
\description{
Function to update the extraction rates and input from processing after
each iteration of Aupus.  This function takes the quantity of a commodity
that was allocated to Reemployment Same Sector in one step and passes that
value on to the edges (to be passed to the children).  Additionally, the
extraction rates computed by the parents are passed on to the children.
}

