% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/updateInputFromProcessing.R
\name{updateInputFromProcessing}
\alias{updateInputFromProcessing}
\title{Update Input from Processing}
\usage{
updateInputFromProcessing(nodes, edges, aupusParam)
}
\arguments{
\item{nodes}{The nodes data returned by the function buildNodes.}

\item{edges}{The edge data returned by the function buildEdges.}

\item{aupusParam}{A list of running parameters to be used in pulling the data.
Typically, this is generated from getAupusParameter (see that function for
a description of the required elements).}
}
\value{
No values are returned, but instead the passed nodes dataset is
updated.
}
\description{
The edge dataset contains information about the quantities of a particular
item flowing from a particular parent to a particular child.  This function
aggregates the edge dataset, grouping by child, and replaces element 31
(Actual Producing Factor) of the child with the inputs from all the
parents.
}

